# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tutor Reservation API",
        "description": "RESTful API for tutor reservations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main server"
        }
    ],
    "paths": {
        "/api/admin/users": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a user",
                "description": "Creates an admin, tutor, or student. Only users with role_id 1 can use this endpoint.",
                "operationId": "e8099f9cb4d98a9dcb191e75b1453c9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_names",
                                    "last_names",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role_id"
                                ],
                                "properties": {
                                    "first_names": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "last_names": {
                                        "type": "string",
                                        "example": "Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.perez@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Admin12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Admin12345"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "enum": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "hourly_rate": {
                                        "type": "integer",
                                        "example": 40000,
                                        "nullable": true,
                                        "maximum": 100000,
                                        "minimum": 30000
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "3001234567",
                                        "nullable": true
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "photos/users/carlos.jpg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not an admin"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer token generated by Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}